+++ src/main/java/com/bank/accounts/dto/AccountDto.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,5 +14,5 @@ public class AccountDto {
     private Long id;
     private String nrb;
     private String currency;
-    private Double availableFunds;
+    private BigDecimal availableFunds;
 }
